+++ src/ocean/sys/CpuAffinity.d
@@ -8,6 +8,10 @@
 
     Functions to set a process' CPU affinity.
 
+    This module uses the GNU API and contains declarations and functionality
+    found in <sched.h> on GNU/Linux. See
+    http://www.gnu.org/software/libc/manual/html_node/CPU-Affinity.html
+
 *******************************************************************************/
 
 module ocean.sys.CpuAffinity;
@@ -60,6 +64,9 @@ private extern ( C )
 
 public struct CpuAffinity
 {
+    import ocean.sys.ErrnoException;
+    import tango.stdc.errno: EINVAL;
+
 static:
 
     /***************************************************************************
@@ -69,19 +76,45 @@ static:
         Params:
             cpu = index of cpu to run process on
 
-        Returns:
-            true on success, false on failure (errno will be set)
+        Throws:
+            ErrnoException on failure. Possible errors:
+            - EINVAL: The processor is not currently physically on the system
+                and permitted to the process according to any restrictions that
+                may be imposed by the "cpuset" mechanism described in cpuset(7).
+            - EPERM: The calling process does not have appropriate privileges.
 
     ***************************************************************************/
 
-    public bool set ( uint cpu )
+    public void set ( uint cpu )
     {
-        cpu_set_t cpu_set;
-        CPU_SET(cast(__cpu_mask)cpu, cpu_set);
+        try
+        {
+            cpu_set_t cpu_set;
+            CPU_SET(cast(__cpu_mask)cpu, cpu_set);
 
-        const pid_t pid = 0; // 0 := calling process
-        auto ret = sched_setaffinity(pid, cpu_set_t.sizeof, &cpu_set);
-        return ret == 0;
+            const pid_t pid = 0; // 0 := calling process
+            if (sched_setaffinity(pid, cpu_set_t.sizeof, &cpu_set))
+            {
+                throw (new ErrnoException).useGlobalErrno("sched_setaffinity");
+            }
+        }
+        catch (ErrnoException e)
+        {
+            /*
+             * Add a sensible error message for EINVAL because this error can be
+             * caused by a bad parameter value in the config.ini, and the
+             * standard message "sched_setaffinity: Invalid argument" may appear
+             * cryptic to the user.
+             */
+
+            if (e.errorNumber == EINVAL)
+            {
+                e.append(" - probably attempted to set the CPU affinity to a " ~
+                         "CPU that doesn't exist in this system.");
+            }
+
+            throw e;
+        }
     }
 
 
@@ -154,14 +187,24 @@ static:
             cpu = cpu index
             set = cpu set
 
-        Returns:
-            cpu set with bit mask with the indexed cpu set to 1
+        Throws:
+            ErrnoException (EINVAL) if cpu is too high to fit in set.
 
     ***************************************************************************/
 
     private void CPU_SET ( uint cpu, ref cpu_set_t set )
     {
-        set.__bits[CPUELT(cpu)] |= CPUMASK(cpu);
+        auto i = CPUELT(cpu);
+
+        if (i < set.__bits.length)
+        {
+            set.__bits[i] |= CPUMASK(cpu);
+        }
+        else
+        {
+            throw (new ErrnoException).set(EINVAL, "CPU_SET")
+                                      .append(" - CPU index too high");
+        }
     }
 }